control SeqWeightedLoss class with beta parameter
This commit updates "SeqWeightedLoss" in the "papote/data_utils.py"
file. This class is used for calculating the weighted cross-entropy loss
during training. The beta parameter is added to control the exponential
average of the loss. This change is made to improve the model's
performance by using a more sophisticated loss function that downweighs
the first tokens (which are very hard to predict), and upweighs the last
tokens (easier to predict) so that the model focuses more on long term
semantics rather than guessing.