Add TopkAccAvg callback to train.py for tracking token accuracy during training
This commit adds the TopkAccAvg callback from tcb to train.py, which allows tracking of accuracy during training. This callback is used to monitor the performance of the model and provide a more comprehensive evaluation of the training process by monitoring the token accuracy which might not agree with token loss. We track to see if there is a fundamental flaw in using loss.